Move pybind11/compat to pybind11/conduit as suggested by @henryiii:
https://github.com/pybind/pybind11/pull/5375#pullrequestreview-2329006001
