+++ include/pybind11/detail/common.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include <pybind11/compat/wrap_include_python_h.h>
+#include <pybind11/conduit/wrap_include_python_h.h>
 #if PY_VERSION_HEX < 0x03080000
 #    error "PYTHON < 3.8 IS UNSUPPORTED. pybind11 v2.13 was the last to support Python 3.7."
 #endif
