+++ tests/extra_python_package/test_files.py
@@ -50,11 +50,11 @@ main_headers = {
     "include/pybind11/warnings.h",
 }
 
-compat_headers = {
-    "include/pybind11/compat/README.txt",
-    "include/pybind11/compat/pybind11_conduit_v1.h",
-    "include/pybind11/compat/pybind11_platform_abi_id.h",
-    "include/pybind11/compat/wrap_include_python_h.h",
+conduit_headers = {
+    "include/pybind11/conduit/README.txt",
+    "include/pybind11/conduit/pybind11_conduit_v1.h",
+    "include/pybind11/conduit/pybind11_platform_abi_id.h",
+    "include/pybind11/conduit/wrap_include_python_h.h",
 }
 
 detail_headers = {
@@ -104,7 +104,7 @@ py_files = {
     "setup_helpers.py",
 }
 
-headers = main_headers | compat_headers | detail_headers | eigen_headers | stl_headers
+headers = main_headers | conduit_headers | detail_headers | eigen_headers | stl_headers
 src_files = headers | cmake_files | pkgconfig_files
 all_files = src_files | py_files
 
@@ -113,7 +113,7 @@ sdist_files = {
     "pybind11",
     "pybind11/include",
     "pybind11/include/pybind11",
-    "pybind11/include/pybind11/compat",
+    "pybind11/include/pybind11/conduit",
     "pybind11/include/pybind11/detail",
     "pybind11/include/pybind11/eigen",
     "pybind11/include/pybind11/stl",
