+++ tools/setup_global.py.in
@@ -26,14 +26,14 @@ class InstallHeadersNested(install_headers):
 
 
 main_headers = glob.glob("pybind11/include/pybind11/*.h")
-compat_headers = sum([glob.glob(f"pybind11/include/pybind11/compat/*.{ext}")
-                      for ext in ("h", "txt")], [])
+conduit_headers = sum([glob.glob(f"pybind11/include/pybind11/conduit/*.{ext}")
+                       for ext in ("h", "txt")], [])
 detail_headers = glob.glob("pybind11/include/pybind11/detail/*.h")
 eigen_headers = glob.glob("pybind11/include/pybind11/eigen/*.h")
 stl_headers = glob.glob("pybind11/include/pybind11/stl/*.h")
 cmake_files = glob.glob("pybind11/share/cmake/pybind11/*.cmake")
 pkgconfig_files = glob.glob("pybind11/share/pkgconfig/*.pc")
-headers = main_headers + compat_headers + detail_headers + eigen_headers + stl_headers
+headers = main_headers + conduit_headers + detail_headers + eigen_headers + stl_headers
 
 cmdclass = {"install_headers": InstallHeadersNested}
 $extra_cmd
@@ -57,7 +57,7 @@ setup(
         (base + "share/cmake/pybind11", cmake_files),
         (base + "share/pkgconfig", pkgconfig_files),
         (base + "include/pybind11", main_headers),
-        (base + "include/pybind11/compat", compat_headers),
+        (base + "include/pybind11/conduit", conduit_headers),
         (base + "include/pybind11/detail", detail_headers),
         (base + "include/pybind11/eigen", eigen_headers),
         (base + "include/pybind11/stl", stl_headers),
