+++ tools/setup_main.py.in
@@ -14,7 +14,7 @@ setup(
     packages=[
         "pybind11",
         "pybind11.include.pybind11",
-        "pybind11.include.pybind11.compat",
+        "pybind11.include.pybind11.conduit",
         "pybind11.include.pybind11.detail",
         "pybind11.include.pybind11.eigen",
         "pybind11.include.pybind11.stl",
@@ -24,7 +24,7 @@ setup(
     package_data={
         "pybind11": ["py.typed"],
         "pybind11.include.pybind11": ["*.h"],
-        "pybind11.include.pybind11.compat": ["*.h", "*.txt"],
+        "pybind11.include.pybind11.conduit": ["*.h", "*.txt"],
         "pybind11.include.pybind11.detail": ["*.h"],
         "pybind11.include.pybind11.eigen": ["*.h"],
         "pybind11.include.pybind11.stl": ["*.h"],
